Restore correct environment
A previous PR made the usage and structure of `structured_env` simpler.
However, the effects wasn't the expected one, as the `PATH` became unset
in the development shell. It seems the right fix would be to get rid of
`env` altogether in the builders, and delegate that to `NickelDerivation`.

Unfortunately, `NickelDerivation` ends up being applied several times, and
it's not idempotent (Nickel currently refuses to merge arrays together,
hence the computed value for `env` fails to merge with itself).

We fix the previous bug, and moved the computation of `env` to
`BashShell` in the meantime (waiting for Nickel to be smarter about
merging identical values).