Only coerce the paths to strings if they are in the store
When passing data from Nix to Nickel, we don't blindly `toJSON` the
values of type `path` as that would copy them to the store, which is
redundant for store paths.
However, `path`s aren't necessarily _store_ paths, and if they aren't we
probably want them to be copied to the store (at least as a first
approximation). So add a quick and dirty filtering to only preprocess
the actual store paths.