Radically cleanup the templates
Remove all the templates, except for one.
This makes the repository quite simpler, and emphasises the fact that we
want something that just works™

We'll certainly want to add some funkier tests, but in a more principled
way, and that doesn't leak into the end-user interface.

Also removes the example for the same reason