Merge the different kind of Nixel _things_
Currently we can have a `package.ncl` defining the package we want, and
a `dev-shell.ncl` defining the shells.

Unify both in a `project.ncl` that can export both shells and packages.

We can't just make the flake output be the result of the Nickel call
because we need to be able to access `apps.regenerate-lockfile` even
when we can't evaluate the Nickel expression.
Would be nice to find a way around that.