Autogenerate nickel.lock.ncl
Re-generate expected contents of nickel.lock.ncl on every importNcl call,
check if it matches one in the sources and generate a new source tree with
new nickel.lock.ncl if it doesn't. Note that expected lockfile contents
depend only on flake parameters. Also note that when lockfile contents
don't match the expected value, we'll make a full copy of the source tree
during evalutaion of Nickel file.

Fixes #105
Fixes #116