Add the nickel LSP to the dev shell
The dependency to the lsp was accidentally removed by #139 which changed
the dependency from `nickel#default` (a buildenv of both the lsp and the
cli) to `nickel#nickel-lang-cli`.

Now that we're using organist, we can actually have the `build` shell
depend only on the cli and the dev one add the lsp.