Namespace the Nix-related things
Organist isn't _just_ a Nix framework, so we want to make sure that it
is possible to add stuff that's not strictly Nix-related to it, and have
it feel first-class.
So move most of the Nix-related stuff in its own namespace to clear up
some real estate for the rest.