haskell: Make ormolu the same version as the one used in hls
There can be a discrepancy between `haskellPackages.ormolu` and the
`ormolu` used by `haskellPackages.haskell-language-server`. Having that
means that there can be a discrepancy between how the lsp will format
code and how the cli tool will.

Fix that by taking (by default) ormolu from the closure of
`haskell-language-server` rather than directly from the package set.

The evil genius of this is to be credited to @sir4ur0n, I merely ported
it to Organist.