Fix nls in the tests
`nls --version` isn't a valid command, causing the test suite to fail.

Because `nls` is supposed to be in every shell, the previous
shell-by-shell overriding of the command used to make sure that the
command is here can't really work, so change it to dispatch based just on the
dependency name.

That way we can just say that for every shell, `nls` should be tested
with `nls --help` rather than `nls --version`